refactor(providers): move contracts to index.ts
