--- src/providers/contracts/ProviderOptions.ts
@@ -1,6 +1,0 @@
-export interface ProviderOptions {
-  scheme: string
-  clientSecret: string | number
-}
-
-export default ProviderOptions
