+++ src/providers/facebook/index.ts
@@ -1,5 +1,5 @@
 import { assignDefaults } from '../../utils/provider'
-import ProviderPartialOptions from '../contracts/ProviderPartialOptions'
+import { ProviderPartialOptions } from '../index'
 import FacebookProviderOptions from './contracts/FacebookProviderOptions'
 
 export default function facebook(
