+++ src/providers/index.ts
@@ -1,11 +1,13 @@
-import { PartialExcept } from '../../types/utils'
-import SchemeOptions from '../../schemes/contracts/SchemeOptions'
-import ProviderOptions from './ProviderOptions'
+import SchemeOptions from '../schemes/contracts/SchemeOptions'
+import { PartialExcept } from '../types/utils'
+
+export interface ProviderOptions {
+  scheme: string
+  clientSecret: string | number
+}
 
 type ProviderOptionsKeys = Exclude<keyof ProviderOptions, 'clientSecret'>
 
 export type ProviderPartialOptions<
   Options extends ProviderOptions & SchemeOptions
 > = PartialExcept<Options, ProviderOptionsKeys>
-
-export default ProviderPartialOptions
