+++ src/providers/laravel/passport/contracts/LaravelPassportPasswordProviderOptions.ts
@@ -1,5 +1,5 @@
-import ProviderOptions from '../../../contracts/ProviderOptions'
 import RefreshSchemeOptions from '../../../../schemes/refresh/contracts/RefreshSchemeOptions'
+import { ProviderOptions } from '../../../index'
 
 export interface LaravelPassportPasswordProviderOptions
   extends ProviderOptions,
